KE-28303: fix streaming job list display error when project is empty (#24197)

+++ src/server-base/src/main/java/io/kyligence/kap/rest/response/StreamingJobResponse.java
@@ -56,5 +56,5 @@ public class StreamingJobResponse extends StreamingJobMeta{
     private Long lastStatusDuration;
 
     @JsonProperty("model_indexes")
-    private Long modelIndexes;
+    private Integer modelIndexes;
 }
